+++ cmd/holotreeHash.go
@@ -2,8 +2,12 @@ package cmd
 
 import (
 	"fmt"
+	"io"
+	"os"
+	"strings"
 
 	"github.com/robocorp/rcc/common"
+	"github.com/robocorp/rcc/conda"
 	"github.com/robocorp/rcc/htfs"
 	"github.com/robocorp/rcc/operations"
 	"github.com/robocorp/rcc/pretty"
@@ -12,7 +16,8 @@ import (
 )
 
 var (
-	holotreeShowBlueprint bool
+	holotreeShowBlueprint      bool
+	holotreeReadInputFromStdin bool
 )
 
 var holotreeHashCmd = &cobra.Command{
@@ -24,8 +29,31 @@ var holotreeHashCmd = &cobra.Command{
 		if common.DebugFlag() {
 			defer common.Stopwatch("Conda YAML hash calculation lasted").Report()
 		}
-		_, holotreeBlueprint, err := htfs.ComposeFinalBlueprint(args, "", common.DevDependencies)
-		pretty.Guard(err == nil, 1, "Blueprint calculation failed: %v", err)
+
+		var err error
+		if holotreeReadInputFromStdin {
+			// Now, read from stdin until EOF and use that as the input to calculate the blueprint hash.
+			data, err := io.ReadAll(os.Stdin)
+			pretty.Guard(err == nil, 1, "Failed to read from stdin: %v", err)
+
+			if len(args) != 1 {
+				common.Fatal("When reading from stdin, the target file path must be provided as the first (and only) argument.", nil)
+			}
+
+			filename := args[0]
+			if !strings.HasSuffix(filename, ".yaml") {
+				common.Fatal("When reading from stdin, the target file path must be provided as the first (and only) argument and it must be a `.yaml` file.", nil)
+			}
+
+			environment, err := conda.ReadPackageCondaYamlFromContents(data, filename, common.DevDependencies)
+			pretty.Guard(err == nil, 1, "Failed to read from stdin: %v (filename: %q)", err, filename)
+
+			holotreeBlueprint, err = htfs.BlueprintFromEnvironment(environment)
+			pretty.Guard(err == nil, 1, "Failed to calculate blueprint from environment: %v", err)
+		} else {
+			_, holotreeBlueprint, err = htfs.ComposeFinalBlueprint(args, "", common.DevDependencies)
+			pretty.Guard(err == nil, 1, "Blueprint calculation failed: %v", err)
+		}
 		hash := common.BlueprintHash(holotreeBlueprint)
 
 		common.Log("Blueprint hash for %v is %v.", args, hash)
@@ -59,4 +87,5 @@ func init() {
 	holotreeHashCmd.Flags().BoolVarP(&common.DevDependencies, "devdeps", "", false, "Include dev-dependencies from the `package.yaml` when calculating the hash (only valid when dealing with a `package.yaml` file).")
 	holotreeHashCmd.Flags().BoolVarP(&holotreeJson, "json", "j", false, "Show environment as JSON.")
 	holotreeHashCmd.Flags().BoolVarP(&holotreeShowBlueprint, "show-blueprint", "", false, "Show full blueprint, not just hash.")
+	holotreeHashCmd.Flags().BoolVarP(&holotreeReadInputFromStdin, "stdin", "", false, "Read the conda.yaml/package.yaml contents from stdin.")
 }